services/gcs_cold_loader/requirements.txt
google-cloud-pubsub
google-cloud-storage
fastavro
jsonschema
